Refactors and Added scanner logic
Refactored Scanner.dart to be called RootApp.dart. Renamed HomePage.dart to be called ScannerWidget.dart. Added parsing of URL logic to when the URL is scanned.Added a little bit of error handling in the GUI for if a QR code is invalid. Lastly, added navigation to the RootApp.dart so what we can use Navigator class to handle navigation later.