Auto log in (#9)
* Refactored AuthSystem so it can pull Asynchronosly to allow automatic log in. Also added dependency injection inside for anywhere the authetnicator is being used. Files to review: AppRouter.dart, Authenticator.dart, AuthSystem.dart, Loading.dart, Login.dart, RootApp.dart. To test auto login launch and log in like normal. Then go to /loadingscreen and it should automatically take you to the scanner page.

* Fixed changes

Added Settings.dart. Also added settings route to AppRouter.dart. This allows the user to sign out from the settings page and route them back to the login in screen. Also changed some of the other style of buttons on various page.

* Fixed password characters shown in UI

* Updated .gitignore

* Update .gitignore

* Got rid of.gitignore that was recently add for time being

* Updated UI

Updated the UI Login and Signup to show red colors when something isn't correct

* Updated Loading screen for cleaner animation

* Makes assets transparent

* Refactored Authenticator to use enum types

* Created email verification page. Also doesn't let user in until they have verfied their email

* Email verifcation

Created an email verification page and updated the sign-in process so that it does not allow a user to sign in without verifying their email address.

* Finished email verification

Allows user to verify their email and then handles their login appropriately.

* Fixed unit tests

---------

Co-authored-by: nolan-meyer1 <145584308+nolan-meyer1@users.noreply.github.com>